Release Candidate
I'm considering this a "Release Candidate" as I've finished the URL, grammar, and spelling changes.  At this point there are still empty places in the document that the original author has not filled in and thus I can't consider this a complete 1.0 version until (or if ever) those areas are finished.